+++ src/components/Tasks/tasks.tsx
@@ -1,4 +1,7 @@
-import React from 'react';
+import React, {useEffect} from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import { RootState } from '../../app/rootReducer';
+import {fetchTasksByProject, fetchUsers} from '../../app/tasksReducer';
 import TaskList from './taskList';
 import AddTask from './addTask';
 import TasksStatus from './taskStatus';
@@ -6,22 +9,37 @@ import Container from '@material-ui/core/Container';
 import Card from '@material-ui/core/Card';
 import CardContent from '@material-ui/core/CardContent';
 import CardActions from '@material-ui/core/CardActions';
+import CircularProgress from '@material-ui/core/CircularProgress';
 import '../../styles/tasksStyles.css';
 
 const Tasks = () => {
 
+  const loading = useSelector((state:RootState) => state.tasks.loading)
+  const tasks = useSelector( (state:RootState) => state.tasks.tasks);
+  const project = useSelector( (state:RootState) => state.tasks.projectId);
+  const users = useSelector( (state:RootState) => state.tasks.users)
+  const dispatch = useDispatch();
+
+  useEffect( () => {
+      dispatch(fetchTasksByProject(project));
+      dispatch(fetchUsers())
+  }, [project, dispatch]);
+
   return (
     <React.Fragment>
         <Container>
-          <TasksStatus />
+          {loading ? <CircularProgress className='tasksBox__spinner'/> :
+          (<React.Fragment>
+            <TasksStatus tasks={tasks}/>
           <Card variant="outlined" className='tasksBox'>
             <CardContent>
-              <TaskList />
+              <TaskList tasks={tasks}/>
             </CardContent>
             <CardActions>
-              <AddTask />
+              <AddTask project={project} users={users}/>
             </CardActions>
           </Card>
+          </React.Fragment>)}
         </Container>
     </React.Fragment>
   );
